Extract function to getAllowedPrincipalArns to an aws-iam-utils module.
